requirements.txt
streamlit==1.26.1
gspread==5.7.2
oauth2client==4.1.3
pandas==2.0.3
openai==0.27.8
python-dotenv==1.0.0